// prana-next/src/app/types/BiodiversityRanking.ts
export type BiodiversityRanking = {
    country_biodiversity_rank_id?: string;
    country_name: string;
    biodiversity_rank: number;
    amphibians: number;
    amphibians_rank: number;
    birds: number;
    birds_rank: number;
    fish: number;
    fish_rank: number;
    mammals: number;
    mammals_rank: number;
    reptiles: number;
    reptiles_rank: number;
    plants: number;
    plants_rank: number;
    created?: Date;
    updated?: Date;
};